[lib] Dispatch SET_CONNECTION_ISSUE when keyserver auth returns client_version_unsupported
This helps on `web`, where `VersionHandler` looks at `connectionIssue`s in the `KeyserverStore` to determine whether to show `VersionUnsupportedModal`.

It would be pretty unusual for keyserver auth to be the first place to notice `client_version_unsupported` for a given keyserver, as it would normally be caught before the keyserver is added, or after a session is already established with that keyserver. To test this, I had to disable checks in `AddKeyserver` on `native` so that the keyserver would be added anyways.