[identity] Get and verify initial device list from registration requests
Summary:

Added some helper traits and implemented them for all request types that are registration-related (have initialDeviceList).
This is similiar to the existing `DeviceKeyUploadData` and `DeviceKeyUploadActions` which support common functionality for all request types that have device keys payload.

The trait provides a single function for extracting and verifying the initial device list. The `construct_user_registration_info()` calls it. The device list is then stored for further processing (next diff).

Depends on D12089

Test Plan:

Tested in the next diff when all pieces are joined together.

Reviewers: kamil, will

Subscribers: