[SQLite] recreate table for Outbound P2P messages
Summary:
1. Updating table name to be more readable.
2. Adding column with status.

Note 1:
This table is empty at current state so it's safe to just drop and create.

Note 2:
With current design we allow for `ciphertext` to be empty string. This is case when we need to send something to peer but `olm` session not exists or is malformed. Alternative is making this field nullable, but this will make implementation more complex and will require a lot of additional code (mostly because of [ENG-5719](https://linear.app/comm/issue/ENG-5719/improve-pointers-handling)) so going for empty string when ciphertext is missing.

Note 3:
We need a bit different API to this table - updated in next diffs along with tests.

Depends on D12047

Test Plan:
1. Migration succeeded.
2. Creating DB from scratch works.
3. Tests

Reviewers: tomek, marcin