[SQLite] add query to set ciphertext for outbound P2P message
Summary:
After session creation we want to encrypt the message already addressed for other peer, we need a method for it.

This method should be called in the same SQLite transaction as encryption, this is handled later in the stack.

Using hardcoded constants in code - for later usage on the JS side I'll introduce type to make it safer,

Depends on D12069

Test Plan: Tests

Reviewers: tomek, marcin

Reviewed By: marcin

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D12070