[SQLite] add query to mark outbound P2P message as sent
Summary:
We want to mark message as sent to avoid sending this again. Attempt to send message should happen only for encrypted messages, and even if marking as sent fails this is not a problem - this message will be sent but message will be discarded due to not being able to decrypt the same message twice.

Using hardcoded constants in code - for later usage on the JS side I'll introduce type to make it safer.

Depends on D12070

Test Plan: Tests

Reviewers: tomek, marcin