[lib][web][native] Use httpMessageUserInfosHandler
Summary:
issue: [[https://linear.app/comm/issue/ENG-7821/call-id-extraction-function-in-a-socket-handler-and-a-wrapper-on | ENG-7821]]
As explained in the issue, when I tried using `endpointValidators` in `callSingleKeyserverEndpoint` it created import cycels which are hard to fix. An explenation can be found [[https://linear.app/comm/issue/ENG-8080/create-a-mapping-between-endpoints-and-output-validators-in-lib#comment-b6acb75d | here]].
The bottom line is that I cannot import `lib/types/validators/endpoint-validators.js` from `lib/keyserver-conn/call-single-keyserver-endpoint.js`. So my workaroud is for `lib/keyserver-conn/call-single-keyserver-endpoint.js` to expose a function that lets some other place in the code insert the handler to it.

Test Plan:
Tested that both web and native apps can be run and no errors are thrown. Tested that the handler correctly extracts user ids from http messages.
Reviewers:
tomek, kamil
Subscribers: