[native] Use getMessageForException in registrationServerCall
Summary:
Noticed a couple places where we were accessing `e.message` directly, which can be dangerous if a primitive is thrown.

Depends on D12083

Test Plan: I didn't test this beyond how I tested the rest of the stack. Based on my familiarity with this code, it should be a safe change

Reviewers: inka

Reviewed By: inka

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D12084