[native] Handle client version unsupported errors in registrationServerCall
Summary: We were handling these in some places, but not all. This diff makes sure we handle it all places in `registrationServerCall`, which gets us closer to resolving [ENG-8113](https://linear.app/comm/issue/ENG-8113/client-version-errors-during-auth-are-not-handled-in-most-cases).

Depends on D12084

Test Plan:
There are a lot of spots here, and I didn't check all of them. I verified the core parts:

1. That identity service returns `Unsupported version` and keyserver returns `client_version_unsupported` when the client version is outdated
2. That the `Alert`s showed correctly
3. That `onAlertAcknowledged` worked

Reviewers: inka

Subscribers: