[web] Keep sign in button the same size when loading
Summary:
Noticed that the button shifted in height when loading. Prior to this diff, the button was increasing in height when loading, due to an explicit `min-height` property that gets set when it is `disabled`. It looks like this was built by an intern several years ago, and the `min-height` was probably added to make sure the button height didn't go to 0 or some small number.

This diff changes it so the height is based on the text, and the spinner is absolutely-positioned over the text. To keep the height the same even when the text is hidden, we use `visibility: hidden` for the text instead of `display: none`.

Depends on D12085

Test Plan: [video]

Reviewers: ginsu

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D12086