[web] Avoid changing height when error occurs in log in screen
Summary: This diff hardcodes the error to have `height: 0`, and increases the margin around `OrBreak` to make up for that. I don't love that asymmetry of this solution when displaying the error, but the extra spacing doesn't look too bad.

`OrBreak` is used in one other place (`SIWELogInForm`), so I included a screenshot of that screen in the Test Plan.

Depends on D12086

Test Plan: [video]
[screenshot]

Reviewers: ginsu

Subscribers: