[native] Stop persisting invite links
Summary: We don't need to persist the links because we're fetching them every time we open the community drawer. We will need to start persisting the links in SQLite after https://linear.app/comm/issue/ENG-3937/distribute-link-changes-through-updates is done.

https://linear.app/comm/issue/ENG-6920/move-invitelinksstore-to-sqlite#comment-734f824b

Test Plan: Run the app and check if a warning about rehydrated keys not expected is shown. Open the drawer - a link should be present for communities that have one. Close and reopen the app - the warning shouldn't appear.

Reviewers: inka