[protos] Add RPC for secondary device logout
Summary:
Addresses [[ https://linear.app/comm/issue/ENG-7886/add-logoutseccondarydevice-rpc | ENG-7886 ]].
Introduced `LogOutSecondaryDevice` RPC which is going to be called by secondary device when user clicks the logout button.
This is part of the "CSAT invalidation flow" described in [[ https://linear.app/comm/issue/ENG-5293/add-support-for-logging-out-of-a-secondary-device | ENG-5293 ]].

In contrast to the existing `LogOutUser` RPC, this one will only clear CSAT and device keys, but won't be modifying the device list.

Depends on D12091

Test Plan: CI. Implementation later in the stack

Reviewers: ashoat!, varun, kamil, will

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D12119