[lib][native][web] Convert QRAuthHandler to a hook
Summary:
This component makes more sense as a hook since it doesn't render anything.

In later diffs I'll refactor it further, and it's helpful to first refactor it to a hook.

Depends on D12110

Test Plan: I tested the flow using staging, with my phone as a primary device and a local simulator as a secondary device

Reviewers: bartek, tomek

Subscribers: