[web] Show VersionUnsupportedModal if QR code auth fails due to unsupported client version
Summary: I originally noticed that the QR code flow didn't handle this error when investigating [ENG-8113](https://linear.app/comm/issue/ENG-8113/client-version-errors-during-auth-are-not-handled-in-most-cases). That led to the further realization that the QR code flow hasn't been kept up-to-date, which was addressed in [ENG-8123](https://linear.app/comm/issue/ENG-8123/add-keyserver-auth-to-qr-code-login-flow). This diff completes the circle and finally adds code to handle the outdated client error specifically.

Depends on D12124

Test Plan: I hardcoded a failure in `keyserver/src/session/version.js` for all web clients and saw this error: [screenshot]

Reviewers: bartek

Subscribers: