[web][native] persist Outbound P2P messages when processing store ops
Summary:
It's important to address messages to the other peers in the same SQLite transaction as updating our own store (we don't have state-sync yet so this allows us to make sure data is consistent).

Later on, this data will be encrypted and sent through Tunnelbroker (D12132).

Depends on D12129

Test Plan:
Make sure messages are persisted (added some mock and read it later).
Tested end-to-end in: D12133

Reviewers: tomek, bartek

Reviewed By: bartek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D12130