[web] Retrieve username on secondary device login
Summary:

Addresses [[ https://linear.app/comm/issue/ENG-7215/retrieve-username-when-doing-qr-code-auth | ENG-7215 ]] on web.
Now we have `username` in RCPs response, so we can simply pass it to `identityAuthResult`, replacing the empty string placeholder.

Depends on D12134.

Test Plan:

Console.log the `identityAuthResult` in the `uploadKeysForRegisteredDeviceAndLogIn()` and confirmed the username was there during QR login.

Reviewers: inka, varun, will

Subscribers: