[native] Persist only lastUserInteractionCalendar from the entry store
Summary:
We keep this field in redux-persist. `EntryInfos` are stored in SQLite, and `daysToEntries` are computed based on the entries.

Depends on D12169

https://linear.app/comm/issue/ENG-3486/move-entrystore-to-sqlite

Note: this diff should be landed when we are sure that the migration to SQLite was successful

Test Plan: Added some console logs and checked if `lastUserInteractionCalendar` is persisted and rehydrated.

Reviewers: kamil, marcin

Reviewed By: marcin

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D12153