[lib] Update delete keyserver, delete thread, leave thread and set new session actions with the ops approach
Summary:
These actions override the store with a modified one - replicate the logic for the ops.

Depends on D12157

https://linear.app/comm/issue/ENG-3486/move-entrystore-to-sqlite

Test Plan:
Be on the latest diff if this stack to ensure there is no mechanism from the keyserver that would aid entry persistence. After testing each actions kill the app and reopen it to ensure that changes are correctly persisted.

1. To test delete keyserver, delete thread, leave thread by: create community, then a chat in the community, then the entry for the chat and finally deleted the chat. Ensure that it disappeared from SQLite.
2. To test set new session: TODO

Reviewers: kamil, marcin

Reviewed By: marcin

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D12156