[native] Avoid calling our custom Alert from reducers
Summary:
The test plan in D12124 points out an issue where our custom `Alert` is called from a reducer, but I didn't catch all of these cases.

The code in `onStateDifference` is called from reducers as well, so we should use the stock React Native `Alert` here as well.

Test Plan: I have a repro of this locally, and I was able to confirm that the error no longer occurs when the `Alert` is triggered

Reviewers: ginsu, marcin

Reviewed By: ginsu, marcin

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D12159