[keyserver] call uploadSecondaryDeviceKeysAndLogIn
Summary:
Depends on D12155

after device list has been updated, secondary device should get a nonce from the keyserver, sign it, and then log in to identity

i also took this chance to introduce a helper function to get the device key upload

Test Plan:
successfully logged in to identity and got an access token

checked that the device key upload contents were all present and looked correct. the one-time keys are marked as published in the helper function, but i will have to mark the prekeys as published after the identity service call in the next diff.

Reviewers: bartek, marcin, kamil, ashoat

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D12161