[web][keyserver] Use entries from the initial state only if they are missing from the store
Summary:
This is the last step of the migration - we should no longer use the values returned from the initial Redux state.

Depends on D12153

https://linear.app/comm/issue/ENG-3486/move-entrystore-to-sqlite

Test Plan: Closed and opened the app (both native and web) - the entries aren't sent from the keyserver so doing this verified if they are correctly persisted and rehydrated.

Reviewers: kamil, marcin

Reviewed By: marcin

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D12170