[native] Inline siweButton in LoggedOutModal
Summary:
As a first step towards replacing the Reanimated 1 syntax in `LoggedOutModal` with Reanimated 2 syntax ([ENG-8146](https://linear.app/comm/issue/ENG-8146/convert-loggedoutmodal-to-reanimated-2-syntax)), I'm converting the class component to a function component.

Not clear why this was defined separately. Maybe it was used in multiple places previously?

Doing this as a separate diff because it's a simple move, to make later diffs easier to review.

Depends on D12192

Test Plan:
I tested this diff stack by playing around with `LoggedOutModal` on my iOS simulator in my local environment:

1. I navigated to `LogInPanel` and back
2. I navigated to `LegacyRegisterPanel` and back
3. I navigated to `FullscreenSIWEPanel` and back
4. I navigated to `QRCodeSignInNavigator` and back
5. I navigated to `Registration` and back

Reviewers: inka

Reviewed By: inka

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D12193