[native] Reduce indentation of LoggedOutModal
Summary:
As a first step towards replacing the Reanimated 1 syntax in `LoggedOutModal` with Reanimated 2 syntax ([ENG-8146](https://linear.app/comm/issue/ENG-8146/convert-loggedoutmodal-to-reanimated-2-syntax)), I'm converting the class component to a function component.

By moving the `React.memo` wrapper to the end, we're able to reduce the indentation.

This is the final diff in the `LoggedOutModal` -> function component refactor!

Depends on D12197

Test Plan:
I tested this diff stack by playing around with `LoggedOutModal` on my iOS simulator in my local environment:

1. I navigated to `LogInPanel` and back
2. I navigated to `LegacyRegisterPanel` and back
3. I navigated to `FullscreenSIWEPanel` and back
4. I navigated to `QRCodeSignInNavigator` and back
5. I navigated to `Registration` and back

Reviewers: inka