[native] Avoid showing iCloud password manager prompt when dismissing LogInPanel
Summary:
For some reason, this prompt now appears when the `LogInPanel` is dismissed. It seems that switching to Reanimated 2 syntax here changed something about how the iOS password manager logic works with our code. (One big plus is that in `LogInPanel`, the username and password are now autocompleted together, instead of requiring two separate steps – cc @varun, who previously talked about this with me.)

Anyways: it makes sense to show this prompt when the user attempts a login, but not when the user goes back.

To avoid showing the prompt, we can clear the password `TextInput` before it disappears from the screen.

Depends on D12208

Test Plan:
| before | after |
| {F1838435} | {F1838449} |

Reviewers: inka

Reviewed By: inka

Subscribers: varun, tomek

Differential Revision: https://phab.comm.dev/D12209