[keyserver] finish secondary login via qr code
Summary:
after successfully uploading keys and logging in to the identity service, we should mark the prekeys as published and save the access token and user ID returned by the identity service. then, we should close the websocket connection. when it is reopened by the onClose function, it should know that after authenticating with tunnelbroker, it needs to send a `SECONDARY_DEVICE_REGISTRATION_SUCCESS` message to the primary device.

Depends on D12212

Test Plan: successfully added my local keyserver as a secondary device by scanning the qr code from my physical device. verified that the device list was updated in ddb. also received an alert on primary confirming secondary login.

Reviewers: bartek, kamil, ashoat

Reviewed By: ashoat

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D12213