[lib] reverse conditional logic for convertOpsToClientDBOps in message store ops
Summary: In the next diff I will introduce the logic to handle the local field in the message store using the ops approach. As a preliminary step to this introduction, I needed to revers this logic here so that I can add additional conditions in the next diff.

Test Plan: Confirmed that there are no regressions with message store + tested the entire stack at the end of the hackathon

Reviewers:

Subscribers: