[keyserver] Throw an exception when trying to create a DM on a non-auth keyserver
Summary:
Currently the DMs are hosted on the authoritative keyserver and can't be created on non-auth ones.

https://linear.app/comm/issue/ENG-7845/stop-creating-dms-on-non-authoritative-keyservers

Depends on D12228

Test Plan: Tested this by removing authoritative keyserver check when creating threads while setting up a new db - this causes an attempt to create a thread in GENESIS. On authoritative keyserver it works. On a non-auth keyserver an exception is thrown - as expected.

Reviewers: inka, ginsu

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D12229