[lib] Avoid passing userStateBeforeRecovery to invalidateKeyserverSession
Summary:
This is a minor refactor I noticed while working on the prior diff.

There's no reason to pass this around from one callback to the other, since it can be accessed directly on the second callback.

Depends on D12232

Test Plan: Flow, also tested together with the prior diff

Reviewers: inka, tomek

Subscribers: