[native] Only show CommunityDrawerItem for communities
Summary:
We already do this on `web`, since there we render a separate React component for community items (`CommunityDrawerItemCommunity`) than channels (`CommunityDrawerItem`).

On `native` we use the same component. We shouldn't be showing any options for channels, but we currently show an option to admins to manage the roles. Roles should only be managed on a per-community basis, so this diff hides the "three dots" button for all channels. This results in a much cleaner look with fewer "three dots" buttons.

Test Plan: Confirm that the "three dots" button only appears for community roots, and not for channels on my local iOS simulator

Reviewers: inka

Subscribers: