[lib] Allow dash character in invite links
Summary: This resolves [ENG-8263](https://linear.app/comm/issue/ENG-8263/invite-links-with-certain-names-and-hyphens-cannot-be-generated), but most of the work was done in the parent.

Depends on D12241

Test Plan:
Tested invite link flow end-to-end on web by creating an invite link as a keyserver admin, and then opening it in an Incognito tab as another user, and confirming I am able to join

Opted not to test `native` at all, but can test that too if helpful, and especially if we suspect this RegExp is further duplicated elsewhere.

Reviewers: tomek

Subscribers: