[keyserver] Allow staff to use reserved names for invite links
Summary:
This addresses the second part of [ENG-8263](https://linear.app/comm/issue/ENG-8263/invite-links-with-certain-names-and-hyphens-cannot-be-generated): update the keyserver code to allow staff members to set invite links that are reserved.

Depends on D12242

Test Plan:
Tested invite link flow end-to-end on web by creating an invite link named "music" as a keyserver admin, and then opening it in an Incognito tab as another user, and confirming I am able to join. Before modifying `lib/facts/staff.json` to include the local keyserver admin, it did not work. Afterwards it did

I was surprised that the staging blob service did not reject my invite link creation, given that I was using a reserved string. This calls into question whether any changes on the blob service are necessary. Will document this on [ENG-8302](https://linear.app/comm/issue/ENG-8302/stop-checking-if-an-invite-link-contains-a-reserved-username-on-blob).

Reviewers: tomek

Differential Revision: https://phab.comm.dev/D12243