[native] Fix LoggedOutModal buttons not appearing on load in release build
Summary:
This resolves [ENG-8321](https://linear.app/comm/issue/ENG-8321/splash-screen-wont-dismiss-on-release-build).

I made a mistake in D12208. In Reanimated 3 it's okay to pass a Reanimated value directly to a `style` prop of an `Animated.View`, but not in Reanimated 2. I had been reading the Reanimated 3 docs initially, and after learning about this I updated most places but seem to have missed this spot.

In Reanimated 2, all uses of Reanimated values must be wrapped in `useAnimatedStyle`.

Guessing it only shows up in a release build due to some race conditions.

Test Plan:
Confirmed that it resolves the linked Linear issue by compiling a release build to my iOS simulator

Reviewers: varun, will

Subscribers: