[lib][native][web] Reorder results in ThreadPickerModal
Summary:
For the main thread list search, we rely on the [reordering behavior](https://github.com/CommE2E/comm/blob/f8ee46466872e45db17bc98d853e337fad904582/lib/shared/thread-utils.js#L1291-L1311) of `getThreadListSearchResults` to surface the most relevant results first (private and personal threads) when a user searches for a username.

This diff introduces similar functionality for `ThreadPickerModal` on both `native` and `web`. I considered trying to extract the reordering logic, but since it operates on `ChatThreadItem` and I needed to operate on `ThreadInfo`, this would have been messy. Instead I opted to duplicate it.

Note that on `native`, `ThreadList` is used in only two places. One of them is `ThreadPickerModal`, and the other is `ComposeSubchannel`, which doesn't use the searching functionality at all.

Depends on D12250

Test Plan: I tested by connecting my local simulator to the prod keyserver, logging in as a test user, and confirming that when I searched for "ashoat" in the `ThreadPickerModal`, the results were reordered so that the test user's `PERSONAL` thread with me appeared first

Reviewers: inka

Reviewed By: inka

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D12251